Store vector norm with embedding entry
This adds the field `vector_norm`, precalucating the norm when
calculating the embeddings. The reason for this is that we need
the norm when later calculating the cosine similarity; and since
we already have the embedding vector in memory here, it's more
efficient to calculate the norm once and store it, rather than
calculating it again every time we calculate a cosine similarity.

Note, that we do not add it to gemini, considering we do not want
to pay for calculating the embeddings once more. A short script
could probably do this for gemini as well.

Furthermore, we add a comment on how to run the model from a local
directory.