Add filter, improve plotting and restructure code
Previously, the codebase lacked the functionality to filter ECLASS
definitions. As many definitions hold no semantic value, the computation
of these embedding vectors was therefore meaningless.

This adds a filter to sort out non-semantic ECLASS definitions. These
filters are based on manual analysis of the ECLASS definitions and may
therefore not be exhaustive.

Moreover, this provides improved scripts for computing and visualising
clusters and bar charts.

Finally, the code itself was restructured to improve readability.