Add fuzzy logic, plotting for transitivity and tests, improve code efficiency and plotting
Previously, we didn't implement any of the researched methods of fuzzy
logic. Moreover, the code was messy in some places and the plotting was
not standardised.

This adds important fuzzy logic functions, as well as the corresponding
plotting of transitivity relations via matching scores and new tests for
the fuzzy logic functions.

Furthermore, the efficiency of the matching score loading process has
been improved, and all the plotting functions have been standardised.