Add inaccuracy checker
Previously, the code only checked for missing and structural ECLASS
definitions.

This adds a new interface for checking ECLASS definition inaccuracies.
The idea is to collaboratively add different kind of definition checkers
based on the new interface.

As of now, checks for missing definitions, structural definitions,
missing full stops, missing capital letters at the beginning of a
definition, spelling mistakes, duplicate definitions and watermarks in
definitions were implemented based on the new interface.